Put back DEAUTHFLOOD alert
DEAUTH alert will be sent if more than 10 deauthentications or
disassociations are received in one second as done in the
rudimentary mechanism that was added back in 2002:
https://github.com/kismetwireless/kismet/commit/2a132d45ef602d8bde5fbe1d5df8c1c2d673606a

It should be noted that this detection mechanism works better if the
user sets packet_dedup_size to 0 otherwise most of the flooding packets
will be detected as duplicate and so will be discarded.
I would advise to document this limitation.

Fix #227

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>