kis_datasource.cc: fix error display
Since commit 7ad14b75b7813640ccdd22860a386a9702707fd7, an error is
displayed only the first time, for example when loosing the connection
with a remote capture.

This seems a bit strange so remove this limitation.

Moreover, errors are not displayed after the source is closed since
commit 3bab497c7a70610f0942c14feb2e5cb6ff0d6af4.

This is an issue if the source is closed and then re-opened so remove
also this limitation.

Both changes should fix #198

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>