Cut gps coord. lat/lon to 6 decimals
* More decimals do not mean better accuracy or precision but take space for nothing in csv file

See http://wiki.gis.com/wiki/index.php/Decimal_degrees

The 6th decimal of the gps longitude at the equator means roughly 0.11m (around 4 inches).
This is enough precision given the gps device is not able, most of the time, to deliver a better precision than that, when it's not around the meter.
So it does not make sense to store a position with sub-centimeter precision when the error is almost a meter.

I don't know what the wigle guys think about that...